Refactor
---

Original commit (chain-contracts): 2b762c12e187ba069f7dbbca31fcdedbb21c3385

+++ chain-contracts/loom/contracts/controllers/modules/UserPlanetMapUtil.sol
@@ -39,7 +39,7 @@ contract UserPlanetMapUtil {
 
   function _distanceFromCenter(int16 coordinateQ, int16 coordinateR) private pure returns (uint16) {
     return
-      uint16(_max(_abs(coordinateQ), _abs(coordinateR), _abs(-int24(coordinateQ) - coordinateR)));
+      uint16(_max(_abs(coordinateQ), _abs(coordinateR), _abs(-coordinateQ - coordinateR)));
   }
 
   function _max(uint256 a, uint256 b, uint256 c) private pure returns (uint256) {
